requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
aiofiles==24.1.0

# Supabase
supabase==2.7.4

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Config
pydantic-settings==2.4.0
python-dotenv==1.0.1

# Web Scraping
aiohttp==3.10.5
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0

# PDF
reportlab==4.2.2
Pillow==10.4.0
qrcode==8.0

# Data
pandas==2.2.2
openpyxl==3.1.5
